Module rename: registered_contracts -> contracts
