+++ api/engineapi/api.py
@@ -15,7 +15,7 @@ from .routes.dropper import app as dropper_app
 from .routes.leaderboard import app as leaderboard_app
 from .routes.admin import app as admin_app
 from .routes.play import app as play_app
-from .routes.registered_contracts import app as contracts_app
+from .routes.contracts import app as contracts_app
 from .version import VERSION
 
 logging.basicConfig(level=logging.INFO)
